+++ src/hkimjp/wil2/system.clj
@@ -12,7 +12,7 @@
 (defn start-jetty
   []
   (let [port (parse-long (or (env :port) "3000"))
-        handler (if (= (env :develop) "true")
+        handler (if (env :develop)
                   #'routes/root-handler
                   routes/root-handler)]
     (reset! server (jetty/run-jetty handler {:port port :join? false}))
@@ -24,7 +24,8 @@
     (t/log! :info "server stopped.")))
 
 (defn start-system []
-  (t/log! :info (str "start-system redis " (env :redis) " ds " (env :datascript)))
+  (t/log! :info (str "start-system develop: " (env :develop)))
+  (t/log! :info (str "redis " (env :redis) " ds " (env :datascript)))
   (c/redis-server (env :redis))
   (ds/start-or-restore {:url (env :datascript)})
   (start-jetty))