+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 - rename `login` namespace to `auth`
 - sum received points
 - sum sent points
+- shuffle upload?
+- Todays, uploaded is `filtered uploads`
+
+# 0.3.3 (2025-09-16)
+
+- display `develop` when DEVELOP=true
+- renamed `link` to `hx-link`, `button` to `hx-button` in todays.clj
+- ranamed Todays, "uploaded"  to "upload (filtered)"
 
 
 # 0.3.2 (2025-09-10)